Inline typed test fixtures instead of separate helper files
